requirements.txt
matplotlib==3.6.2
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
scikit_learn==1.2.0
segmentation_models_pytorch==0.3.1
torch==1.13.1
torchmetrics==0.11.0
tqdm==4.64.1
neptune-client==0.16.15

